Move internal representations into proximal.experimental.ir namespace
